Get rid of KHtml
KHtml is long obsolete, a dependency we want to avoid, and suffers
from low-quality text rendering (e.g. from inexact font metrics
calculations).  On the other hand, Qt's text widgets are able to
display a HTML subset which is well suited to display the search
results (though we have to make some compromises for the welcome
page, but that's acceptable) and provide better rendering quality.

Therefore, replace the KHTMLPart used to show the search results
with a QTextBrowser.

Tweak the search result's HTML and CSS to display correctly with
the HTML subset available.  (Due to the search path set for the
QTextBrowser and direct access to Qt's palette some of the
substitutions aren't necessary anymore.  Text size is no longer
set explicitly and can be changed on the fly.)

The welcome and no matches pages are replaced with a similary
looking table-based design (yuck!) with some simplified graphics
to cope with the limited CSS support.  Note that ImageMagick is
required at build time to generate the images.